test/web-platform-tests/to-upstream/html/browsers/browsing-the-web/navigating-across-documents/fragment-replace.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>location.replace() with a fragment keeps history intact</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#scroll-to-fragid">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";
setup({ allow_uncaught_exception: false });

async_test(t => {
  // get three urls in the history
  window.location.assign("#apple");
  window.location.assign("#bannana");
  window.location.assign("#cat");
  assert_equals(window.location.hash, "#cat", "Before");

  // go to middle history entry
  window.history.back();
  assert_equals(window.location.hash, "#cat", "Back is async");
  t.step_timeout(() => {
    assert_equals(window.location.hash, "#cat", "Back has a second async step");
    t.step_timeout(() => {
      assert_equals(window.location.hash, "#bannana", "After Back");
      const historyLengthBefore = window.history.length;
      // replace the middle entry
      window.location.replace("#zebra");
      assert_equals(window.location.hash, "#zebra", "After Replace");

      // go forward, spec says this is a no-op because replacing with `#zebra` cleared `#cat`
      // from the history. Actual result is that the url updates to `#cat` in real browsers.
      // See https://github.com/whatwg/html/issues/2796
      history.forward();
      assert_equals(window.location.hash, "#zebra", "Forward is async");
      t.step_timeout(() => {
        assert_equals(window.location.hash, "#zebra", "Forward has a second async step");
        t.step_timeout(() => {
          assert_equals(window.location.hash, "#cat", "After Forward");
          assert_equals(window.history.length, historyLengthBefore, "replacing does not change history length");
          t.done();
        }, 0);
      }, 0);
    }, 0);
  }, 0);

  // assert_equals(history.length === 1, undefined, "Before");


  // window.location.href = "javascript:window.clicked = true;";

  // assert_equals(window.clicked, undefined, "After");

  // t.step_timeout(() => {
  //   assert_equals(window.clicked, true, "After plus a queued task");
  //   t.done();
  // }, 0);
}, "simple case");
</script>
